removed dropdown file
