--- src/layouts/DashboardDropdown/DashboardDropdown.scss
@@ -1,24 +1,0 @@
-@import '../../colours.scss';
-
-
-.dashboard-dropdown {
-  width: 100%;
-  background-color: white;
-  box-shadow: 0px 3px 30px rgba(black, 0.2);
-  padding: 1em;
-  border-radius: 5px;
-  height: 100%;
-  li {
-    margin: 0.5em 0;
-  }
-}
-
-.dropdown-link {
-  text-decoration: none;
-  color: $grey5;
-  transition: all 0.2s ease;
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
