+++ src/layouts/NavWidget/NavWidget.scss
@@ -66,6 +66,28 @@
   margin-top: 0;
 }
 
+.dashboard-dropdown {
+  width: 100%;
+  background-color: white;
+  box-shadow: 0px 3px 30px rgba(black, 0.2);
+  padding: 1em;
+  border-radius: 5px;
+  height: 100%;
+  li {
+    margin: 0.5em 0;
+  }
+}
+
+.dropdown-link {
+  text-decoration: none;
+  color: $grey5;
+  transition: all 0.2s ease;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
 @media screen and (min-width: 769px) {
   .topbar-account-widget {
     .dropdown-wrapper {
